// src/data/types.ts
/*
 * This file was generated by a tool.
 * Rerun sql-ts to regenerate this file.
 */
export interface ExamDb {
  id: string;
  title: string;
  schoolName: string;
  instructions: string;
  examDate: Date;
  isDeleted: boolean;
  createdAt: Date;
  updatedAt: Date;
}
export interface ExamQuestionDb {
  examId: string;
  questionId: string;
}
export interface MigrationsDb {
  id?: number;
  name: string;
  runOn: Date;
}
export interface QuestionDb {
  id: string;
  topicId: string;
  stage: string;
  difficulty: string;
  marks: number;
  answerSize: number;
  question: string;
  isDeleted: boolean;
  createdAt: Date;
  updatedAt: Date;
}
export interface TopicDb {
  id: string;
  name: string;
  identifier: string;
  isDeleted: boolean;
  createdAt: Date;
  updatedAt: Date;
}